Improve coverage + refactor common module in modor_derive
