+++ crates/modor_derive/src/common/mod.rs
@@ -1,0 +1,4 @@
+pub(crate) mod generation;
+pub(crate) mod idents;
+pub(crate) mod lifetimes;
+pub(crate) mod tuples;
